iot-layer/iot-sensor-agent/requirements.txt
# MQTT Client
paho-mqtt>=1.6.1

# Environment Variables
python-dotenv>=1.0.0

# Data Validation
pydantic>=2.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Type Checking (optional, for development)
mypy>=1.0.0
types-python-dateutil>=2.8.0
